clean up

+++ flyway-core/src/main/java/org/flywaydb/core/internal/dbsupport/sqlite/SQLiteTable.java
@@ -23,12 +23,20 @@ import org.flywaydb.core.internal.util.logging.Log;
 import org.flywaydb.core.internal.util.logging.LogFactory;
 
 import java.sql.SQLException;
+import java.util.Collection;
+import java.util.Collections;
 
 /**
  * SQLite-specific table.
  */
 public class SQLiteTable extends Table {
     private static final Log LOG = LogFactory.getLog(SQLiteTable.class);
+
+    /** SQLite system tables are undroppable. */
+    private static final Collection<String> SYSTEM_TABLES =
+        Collections.singleton("sqlite_sequence");
+
+    private final boolean undroppable;
 
     /**
      * Creates a new SQLite table.
@@ -40,11 +48,16 @@ public class SQLiteTable extends Table {
      */
     public SQLiteTable(JdbcTemplate jdbcTemplate, DbSupport dbSupport, Schema schema, String name) {
         super(jdbcTemplate, dbSupport, schema, name);
+        undroppable = SYSTEM_TABLES.contains(name);
     }
 
     @Override
     protected void doDrop() throws SQLException {
-        jdbcTemplate.execute("DROP TABLE " + dbSupport.quote(schema.getName(), name));
+        if (undroppable) {
+            LOG.debug("SQLite system table " + this + " cannot be dropped. Ignoring.");
+        } else {
+            jdbcTemplate.execute("DROP TABLE " + dbSupport.quote(schema.getName(), name));
+        }
     }
 
     @Override
